Rename test classes and tests in jaxrs* modules
