+++ http/jaxrs-reactive/src/test/java/io/quarkus/ts/http/jaxrs/reactive/OpenShiftRESTEasyReactiveMultipartIT.java
@@ -3,5 +3,5 @@ package io.quarkus.ts.http.jaxrs.reactive;
 import io.quarkus.test.scenarios.OpenShiftScenario;
 
 @OpenShiftScenario
-public class OpenShiftMultipartResourceIT extends MultipartResourceIT {
+public class OpenShiftRESTEasyReactiveMultipartIT extends RESTEasyReactiveMultipartIT {
 }
