+++ http/jaxrs-reactive/src/test/java/io/quarkus/ts/http/jaxrs/reactive/RESTEasyReactiveMultipartIT.java
@@ -1,4 +1,4 @@
-package io.quarkus.ts.http.jaxrs;
+package io.quarkus.ts.http.jaxrs.reactive;
 
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.equalTo;
@@ -13,6 +13,7 @@ import javax.ws.rs.core.MediaType;
 
 import org.apache.commons.io.IOUtils;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 
 import io.quarkus.test.scenarios.QuarkusScenario;
@@ -22,8 +23,9 @@ import io.restassured.http.ContentType;
 import io.restassured.response.ValidatableResponse;
 import io.restassured.specification.MultiPartSpecification;
 
+@Tag("QUARKUS-1075")
 @QuarkusScenario
-public class MultipartResourceIT {
+public class RESTEasyReactiveMultipartIT {
 
     private static final String IMAGE_FILE_NAME = "/quarkus.png";
     private static final String TEXT_WITH_DIACRITICS = "Přikrášlený žloťoučký kůň úpěl ďábelské ódy.";
@@ -33,8 +35,8 @@ public class MultipartResourceIT {
 
     @BeforeAll
     public static void beforeAll() throws IOException {
-        imageFile = new File(MultipartResourceIT.class.getResource(IMAGE_FILE_NAME).getFile());
-        imageBytes = IOUtils.toByteArray(MultipartResourceIT.class.getResourceAsStream(IMAGE_FILE_NAME));
+        imageFile = new File(RESTEasyReactiveMultipartIT.class.getResource(IMAGE_FILE_NAME).getFile());
+        imageBytes = IOUtils.toByteArray(RESTEasyReactiveMultipartIT.class.getResourceAsStream(IMAGE_FILE_NAME));
         new Random().nextBytes(randomBytes);
     }
