+++ http/jaxrs/src/test/java/io/quarkus/ts/http/jaxrs/OpenShiftRESTEasyMultipartIT.java
@@ -3,5 +3,5 @@ package io.quarkus.ts.http.jaxrs;
 import io.quarkus.test.scenarios.OpenShiftScenario;
 
 @OpenShiftScenario
-public class OpenShiftMultipartResourceIT extends MultipartResourceIT {
+public class OpenShiftRESTEasyMultipartIT extends RESTEasyMultipartIT {
 }
