+++ http/jaxrs/src/test/java/io/quarkus/ts/http/jaxrs/RESTEasyMultipartCharsetIT.java
@@ -15,7 +15,7 @@ import io.restassured.http.ContentType;
 import io.restassured.specification.MultiPartSpecification;
 
 @QuarkusScenario
-public class AsciiMultipartResourceIT {
+public class RESTEasyMultipartCharsetIT {
 
     public static final String TEXT_WITH_DIACRITICS = "Přikrášlený žloťoučký kůň úpěl ďábelské ódy.";
     private static final String EXPECTED_ASCII_TEXT = new String(TEXT_WITH_DIACRITICS.getBytes(StandardCharsets.UTF_8),
@@ -25,7 +25,7 @@ public class AsciiMultipartResourceIT {
     static RestService app = new RestService().withProperties("us-asscii.properties");
 
     @Test
-    public void testMultipartText() {
+    public void textInMultipartRequestEncodedWithDefaultCharset() {
         MultiPartSpecification multiPartSpecification = new MultiPartSpecBuilder(TEXT_WITH_DIACRITICS)
                 .controlName("text")
                 .header("Content-Type", "text/plain")
