+++ http/jaxrs/src/test/java/io/quarkus/ts/http/jaxrs/RESTEasyMultipartIT.java
@@ -1,4 +1,4 @@
-package io.quarkus.ts.http.jaxrs.reactive;
+package io.quarkus.ts.http.jaxrs;
 
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.equalTo;
@@ -13,7 +13,6 @@ import javax.ws.rs.core.MediaType;
 
 import org.apache.commons.io.IOUtils;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 
 import io.quarkus.test.scenarios.QuarkusScenario;
@@ -23,9 +22,8 @@ import io.restassured.http.ContentType;
 import io.restassured.response.ValidatableResponse;
 import io.restassured.specification.MultiPartSpecification;
 
-@Tag("QUARKUS-1075")
 @QuarkusScenario
-public class MultipartResourceIT {
+public class RESTEasyMultipartIT {
 
     private static final String IMAGE_FILE_NAME = "/quarkus.png";
     private static final String TEXT_WITH_DIACRITICS = "Přikrášlený žloťoučký kůň úpěl ďábelské ódy.";
@@ -35,8 +33,8 @@ public class MultipartResourceIT {
 
     @BeforeAll
     public static void beforeAll() throws IOException {
-        imageFile = new File(MultipartResourceIT.class.getResource(IMAGE_FILE_NAME).getFile());
-        imageBytes = IOUtils.toByteArray(MultipartResourceIT.class.getResourceAsStream(IMAGE_FILE_NAME));
+        imageFile = new File(RESTEasyMultipartIT.class.getResource(IMAGE_FILE_NAME).getFile());
+        imageBytes = IOUtils.toByteArray(RESTEasyMultipartIT.class.getResourceAsStream(IMAGE_FILE_NAME));
         new Random().nextBytes(randomBytes);
     }
